Command refactor 50% complete: Business logic done, UI logic gone
